0_foundations/requirements.txt
python-dotenv
pymupdf4llm
openai
gradio
